+++ src/Rocketeer/Traits/Task.php
@@ -188,7 +188,7 @@ abstract class Task extends Bash
 	public function share($file)
 	{
 		// Get path to current file and shared file
-		$currentFile = $file;
+		$currentFile = $this->releasesManager->getCurrentReleasePath($file);
 		$sharedFile  = preg_replace('#releases/[0-9]+/#', 'shared/', $currentFile);
 
 		// If no instance of the shared file exists, use current one