+++ src/config/config.php
@@ -66,7 +66,10 @@
 		// A list of folders/file to be shared between releases
 		// Use this to list folders that need to keep their state, like
 		// user uploaded data, file-based databases, etc.
-		'shared' => array(),
+		'shared' => array(
+			'{path.storage}/logs',
+			'{path.storage}/sessions',
+		),
 
 		'permissions' => array(